ai-forecasts/requirements.txt
# Core API & Web
fastapi>=0.100.0
streamlit>=1.29.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Agent Framework
crewai>=0.1.0
openai>=1.0.0
langchain>=0.3.25
langchain-openai>=0.3.21

# Data & Analysis
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
plotly>=5.15.0

# Google News Integration
google-search-results>=2.4.2

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6